refactor: transition to user-focused architecture and add core services
- Update .gitignore and remove ML pycache
- Add User handlers and implement Notification, PDF Export, and Validation services
- Refactor frontend layout and components for user-centric experience
- Update database queries, models, and migrations for menopause focus
- Add comprehensive backend test suite for new services and middlewares

+++ backend/internal/http/handlers/utils.go
@@ -122,12 +122,12 @@ func parseIDParam(c *gin.Context, name string) (int64, error) {
 	return strconv.ParseInt(raw, 10, 64)
 }
 
-func getUserID(c *gin.Context) (int32, error) {
+func getUserID(c *gin.Context) (int64, error) {
 	claims, err := getUserClaims(c)
 	if err != nil {
 		return 0, err
 	}
-	return int32(claims.UserID), nil
+	return int64(claims.UserID), nil
 }
 
 func getUserClaims(c *gin.Context) (middleware.UserClaims, error) {
